refactor: migrate purchased tile index tests to scenario tables (#3856)
Phase-2 slice 3: hoist D1 AC fixtures into test_support scenario rows
and collapse purchased_tile_index_test.dart to table-driven loops.
